Added more converted tflite models as test cases
Converted 10 tflite models to MLIR. Did all the models on the tflite models page except the object detection model (COCO SSD MobileNet v1) because it utilizes an operation unsupported by the converter. The convert prints hex values for some "dense" paramater values that make the files extremely hard to read and too big to load into VSCode, so I replaced the hex values with a short string pointing out the removal of the original hex values.